# postman/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Govia API Docs",
    "version": "1.0.0",
    "description": "Comprehensive REST API & Real-Time Socket documentation for the Govia platform."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API Version 1 Base Path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT Bearer token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "LoginResponseData": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsIn..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsIn..."
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ],
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "User logged in successfully"
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number",
                "example": 1
              },
              "limit": {
                "type": "number",
                "example": 10
              },
              "totalPage": {
                "type": "number",
                "example": 1
              },
              "total": {
                "type": "number",
                "example": 1
              }
            },
            "required": [
              "page",
              "limit",
              "totalPage",
              "total"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/LoginResponseData"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ],
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message description"
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "Email is required"
                }
              },
              "required": [
                "path",
                "message"
              ]
            },
            "example": [
              {
                "path": "email",
                "message": "Email is required"
              }
            ]
          },
          "stack": {
            "type": "string",
            "example": "Error: ...\n    at ..."
          }
        },
        "required": [
          "success",
          "message",
          "errorMessages"
        ]
      },
      "UserResponseData": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "65ab1234567890abcdef1234"
          },
          "name": {
            "type": "string",
            "example": "John Citizen"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "role": {
            "type": "string",
            "example": "CITIZEN"
          },
          "image": {
            "type": "string",
            "example": "https://..."
          },
          "phoneNumber": {
            "type": "string",
            "example": "+1234567890"
          },
          "status": {
            "type": "string",
            "example": "active"
          },
          "isVerified": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "_id",
          "name",
          "email",
          "role"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/user/register": {
      "post": {
        "summary": "Register User",
        "description": "Registers a new user account with role-specific details.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string"
                  },
                  "subRole": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "languagesSpoken": {
                    "type": "string"
                  },
                  "preferredAttorney": {
                    "type": "string"
                  },
                  "preferredBailBondsman": {
                    "type": "string"
                  },
                  "licensedStatesToPractice": {
                    "type": "string"
                  },
                  "barAssociationNumber": {
                    "type": "string"
                  },
                  "lawFirmName": {
                    "type": "string"
                  },
                  "officeName": {
                    "type": "string"
                  },
                  "datePassedTheBar": {
                    "type": "string"
                  },
                  "medicalLicenseNumber": {
                    "type": "string"
                  },
                  "specialization": {
                    "type": "string"
                  },
                  "companyName": {
                    "type": "string"
                  },
                  "businessAddress": {
                    "type": "string"
                  },
                  "badgeNumber": {
                    "type": "string"
                  },
                  "assignedNumber": {
                    "type": "string"
                  },
                  "departmentOrPrecinct": {
                    "type": "string"
                  },
                  "didCarNumberChange": {
                    "type": "string"
                  },
                  "newCarNumber": {
                    "type": "string"
                  },
                  "licenseNumber": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "role",
                  "email",
                  "password"
                ],
                "description": "User registration payload",
                "example": {
                  "name": "John Citizen",
                  "email": "john@example.com",
                  "role": "CITIZEN",
                  "password": "Password123!",
                  "phoneNumber": "+1234567890"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully. OTP sent to email.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string",
                          "example": "User created successfully"
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPage": {
                              "type": "number",
                              "example": 1
                            },
                            "total": {
                              "type": "number",
                              "example": 1
                            }
                          },
                          "required": [
                            "page",
                            "limit",
                            "totalPage",
                            "total"
                          ]
                        },
                        "data": {
                          "nullable": true
                        }
                      },
                      "required": [
                        "success",
                        "statusCode"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request or email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/profile": {
      "get": {
        "summary": "Get User Profile",
        "description": "Retrieves current authenticated user profile details.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string",
                          "example": "Operation successful"
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPage": {
                              "type": "number",
                              "example": 1
                            },
                            "total": {
                              "type": "number",
                              "example": 1
                            }
                          },
                          "required": [
                            "page",
                            "limit",
                            "totalPage",
                            "total"
                          ]
                        },
                        "data": {
                          "$ref": "#/components/schemas/UserResponseData"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update User Profile",
        "description": "Updates profile information and avatar image of the current user.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string",
                          "example": "Operation successful"
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPage": {
                              "type": "number",
                              "example": 1
                            },
                            "total": {
                              "type": "number",
                              "example": 1
                            }
                          },
                          "required": [
                            "page",
                            "limit",
                            "totalPage",
                            "total"
                          ]
                        },
                        "data": {
                          "$ref": "#/components/schemas/UserResponseData"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "Get All Users (Admin)",
        "description": "Admin endpoint to list all users with role filtering and pagination.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string",
                          "example": "Operation successful"
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPage": {
                              "type": "number",
                              "example": 1
                            },
                            "total": {
                              "type": "number",
                              "example": 1
                            }
                          },
                          "required": [
                            "page",
                            "limit",
                            "totalPage",
                            "total"
                          ]
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserResponseData"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}